compose/api/requirements.txt
Flask==0.10.1
flask-restful
mysql-connector==2.2.9
